refactor: move `createContentLink` and `createContentAnchor` to `helper/formatting` module
