+++ module/applications/global/prompt-factory.mjs
@@ -3,11 +3,11 @@ import ActorEd from "../../documents/actor.mjs";
 import ItemEd from "../../documents/item.mjs";
 import LearnableTemplate from "../../data/item/templates/learnable.mjs";
 import DialogEd from "../api/dialog.mjs";
-import { createContentAnchor } from "../../utils.mjs";
 import ChooseAdderSubstitutePrompt from "./choose-adder-substitute.mjs";
 import { getSetting } from "../../settings.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as LEGEND from "../../config/legend.mjs";
+import { createContentAnchor } from "../../helper/formatting.mjs";
 
 const { renderTemplate } = foundry.applications.handlebars;
