+++ module/data/roll/horror-mark.mjs
@@ -1,7 +1,7 @@
 import EdRollOptions from "./common.mjs";
-import { createContentAnchor } from "../../utils.mjs";
 import { getSetting } from "../../settings.mjs";
 import * as MAGIC from "../../config/magic.mjs";
+import { createContentAnchor } from "../../helper/formatting.mjs";
 
 /**
  * @typedef { object } EdHorrorMarkRollOptionsInitializationData
