+++ module/data/roll/initiative.mjs
@@ -1,5 +1,6 @@
 import EdRollOptions from "./common.mjs";
-import { createContentAnchor } from "../../utils.mjs";
+
+import { createContentAnchor } from "../../helper/formatting.mjs";
 
 /**
  * Roll options for initiative rolls.
