+++ module/data/roll/knockdown.mjs
@@ -1,5 +1,5 @@
-import { createContentAnchor } from "../../utils.mjs";
 import EdRollOptions from "./common.mjs";
+import { createContentAnchor } from "../../helper/formatting.mjs";
 
 /**
  * @typedef { object } EdKnockdownRollOptionsInitializationData
